[bootStrapCrossMingwQt5_mxe] Support out-of-tree build, fix option description
The install directory has to stay within MXE target.

+++ bootStrapCrossMingwQt5_mxe.sh
@@ -44,34 +44,40 @@ setupEnv()
     if [ "x$external_libmad" = "x1" ]; then
         export EXTRA_CMAKE_DEFS="-DUSE_EXTERNAL_LIBMAD=true $EXTRA_CMAKE_DEFS"
     fi
-    export TOP=$PWD
+    export BUILDTOP=$PWD
+    if [[ $BUILDTOP = *" "* ]]; then
+        echo "The build directory path \"${BUILDTOP}\" contains one or more spaces."
+        echo "This is unsupported by FFmpeg configure."
+        fail "build prerequisites"
+    fi
+    export SRCTOP=$(cd $(dirname "$0") && pwd)
     export BITS=$bits
     if [ "x$BITS" = "x64" ]; then
         export ARCH="x86_64"
     else
         export ARCH="i686"
     fi
-    export MXE_ROOT=$mxerootdir
+    export MXE_ROOT="$mxerootdir"
     export MXE_TARGET=${ARCH}-w64-mingw32.shared
     export QT_SELECT=5
     if [ "x$author_setup" = "x1" ]; then
         authorSetup
     else
-        export QT_HOME=${MXE_ROOT}/usr/${MXE_TARGET}/qt5
-        export MINGW=${MXE_ROOT}/usr/${MXE_TARGET};
-        export PATH=$PATH:${MXE_ROOT}/usr/bin:${QT_HOME}/bin
-        export TOOLCHAIN_LOCATION=${MXE_ROOT}/usr
-        export SDL2DIR=${MXE_ROOT}/usr/${MXE_TARGET}
-        export PARAL="-j $(nproc)"
+        export MINGW="${MXE_ROOT}/usr/${MXE_TARGET}"
+        export QT_HOME="${MINGW}"/qt5
+        export PATH="$PATH":"${MXE_ROOT}"/usr/bin:"${QT_HOME}"/bin
+        export TOOLCHAIN_LOCATION="${MXE_ROOT}"/usr
+        export SDL2DIR="$MINGW"
+        PARAL="-j $(nproc)"
         if [ "x$debug" != "x1" ]; then
-            export INSTALL_DIR=${MINGW}/out/avidemux
+            export INSTALL_DIR="${MINGW}"/out/avidemux
         else
-            export INSTALL_DIR=${MINGW}/out_debug/avidemux
+            export INSTALL_DIR="${MINGW}"/out_debug/avidemux
         fi
     fi
     export CROSS_PREFIX=$MXE_TARGET
-    export PKG_CONFIG_PATH=$MINGW/lib/pkgconfig
-    export PKG_CONFIG_LIBDIR=$MINGW/lib/pkgconfig
+    export PKG_CONFIG_PATH="${MINGW}"/lib/pkgconfig
+    export PKG_CONFIG_LIBDIR="${MINGW}"/lib/pkgconfig
     export CXXFLAGS="-std=c++11"
     export CROSS_C_COMPILER=gcc
     export CROSS_CXX_COMPILER=g++
@@ -85,54 +91,54 @@ fail()
 
 Process()
 {
-    cd $TOP
-    export BUILDDIR=$1
-    export SOURCEDIR=$2
-    export EXTRA=$3
+    BASE=$1
+    SOURCEDIR=$2
+    EXTRA=$3
     GENERATOR="Unix Makefiles"
     if [ "x$debug" = "x1" ]; then
                 DEBUG="-DVERBOSE=1 -DCMAKE_BUILD_TYPE=Debug"
-                BUILDDIR="${BUILDDIR}_debug"
+                BASE="${BASE}_debug"
                 GENERATOR="CodeBlocks - Unix Makefiles"
     fi
-    echo "Building $BUILDDIR from $SOURCEDIR (Extra=$EXTRA)"
+    BUILDDIR="${PWD}/${BASE}"
+    echo "Building in \"${BUILDDIR}\" from \"${SOURCEDIR}\" with EXTRA=<${EXTRA}>)"
     if [ "x$rebuild" != "x1" ]; then
-        rm -Rf ./"$BUILDDIR"
+        rm -Rf "$BUILDDIR"
     fi
     if [ ! -e "$BUILDDIR" ]; then
-        mkdir "$BUILDDIR" || fail mkdir
+        mkdir "$BUILDDIR" || fail "creating build directory"
     fi
-    cd $BUILDDIR
-    cmake -DCROSS=$MINGW \
+    pushd "$BUILDDIR" > /dev/null
+    cmake -DCROSS="$MINGW" \
     -DCMAKE_SYSTEM_NAME:STRING=Windows \
-    -DCMAKE_FIND_ROOT_PATH=$MINGW \
-    -DTOOLCHAIN_LOCATION=$TOOLCHAIN_LOCATION \
-    -DCMAKE_INSTALL_PREFIX=$INSTALL_DIR \
-    -DCMAKE_CROSS_PREFIX=$CROSS_PREFIX \
-    -DCMAKE_C_COMPILER:STRING=${CROSS_PREFIX}-${CROSS_C_COMPILER} \
-    -DCMAKE_CXX_COMPILER:STRING=${CROSS_PREFIX}-${CROSS_CXX_COMPILER} \
-    -DCMAKE_LINKER:STRING=${CROSS_PREFIX}-ld \
-    -DCMAKE_AR:STRING=${CROSS_PREFIX}-ar \
-    -DCMAKE_RC_COMPILER:STRING=${CROSS_PREFIX}-windres \
-    -DAVIDEMUX_TOP_SOURCE_DIR="$TOP" \
+    -DCMAKE_FIND_ROOT_PATH="$MINGW" \
+    -DTOOLCHAIN_LOCATION="$TOOLCHAIN_LOCATION" \
+    -DCMAKE_INSTALL_PREFIX="$INSTALL_DIR" \
+    -DCMAKE_C_COMPILER:STRING="${CROSS_PREFIX}-${CROSS_C_COMPILER}" \
+    -DCMAKE_CXX_COMPILER:STRING="${CROSS_PREFIX}-${CROSS_CXX_COMPILER}" \
+    -DCMAKE_LINKER:STRING="${CROSS_PREFIX}-ld" \
+    -DCMAKE_AR:STRING="${CROSS_PREFIX}-ar" \
+    -DCMAKE_RC_COMPILER:STRING="${CROSS_PREFIX}-windres" \
+    -DAVIDEMUX_TOP_SOURCE_DIR:STRING="$SRCTOP" \
     ${DEBUG} \
     -G "${GENERATOR}" \
     $EXTRA \
-    $SOURCEDIR || fail cmake
-    make $PARAL >& /tmp/log$BUILDDIR || fail "make, result in /tmp/log$BUILDDIR "
-    make install || fail make_install
+    "$SOURCEDIR" || fail "cmake"
+    make $PARAL >& /tmp/log$BASE || fail "make, result in /tmp/log$BASE"
+    make install || fail "install"
     # Only install  component=dev for dev package
     if [ "x$author_setup" = "x1" ]; then
         DESTDIR=${MINGWDEV} cmake -DCOMPONENT=dev -P cmake_install.cmake || fail make_install_dev
     fi
+    popd > /dev/null
 }
 
 usage()
 {
     echo "Usage: bash $0 [OPTION]"
     echo "  --help                 : Print usage"
-    echo "  --32                   : Build a 32 bit application (default)"
-    echo "  --64                   : Build a 64 bit application"
+    echo "  --32                   : Build a 32 bit application"
+    echo "  --64                   : Build a 64 bit application (default)"
     echo "  --debug                : Switch debugging on"
     echo "  --mxe-root=DIR         : Use MXE installed in DIR (default: ${default_mxerootdir})"
     echo "  --rebuild              : Preserve existing build directories"
@@ -186,38 +192,39 @@ dir_check()
 
 create_release_package()
 {
-    if [ ! -e ${INSTALL_DIR}/avidemux.exe ]; then
+    if [ ! -e "${INSTALL_DIR}"/avidemux.exe ]; then
         echo "No avidemux.exe (${BITS} bit) found in ${INSTALL_DIR}, aborting"
         exit 1
     fi
     echo "Preparing package..."
-    cd $TOP
+    pushd "$BUILDTOP" > /dev/null
     PACKAGE_DIR="packaged_mingw_build_${BUILDDATE}"
     if [ "x$debug" = "x1" ]; then
         PACKAGE_DIR="packaged_mingw_debug_build_${BUILDDATE}"
     fi
-    if [ ! -e $PACKAGE_DIR ]; then
-        mkdir $PACKAGE_DIR
+    PACKAGE_DIR="${BUILDTOP}/${PACKAGE_DIR}"
+    if [ ! -e "$PACKAGE_DIR" ]; then
+        mkdir "$PACKAGE_DIR" || fail "creating package directory"
     fi
-    cp -a $INSTALL_DIR $PACKAGE_DIR
-    cd $PACKAGE_DIR
+    cp -a "$INSTALL_DIR" "$PACKAGE_DIR"
+    cd "$PACKAGE_DIR"
     mv -v avidemux avidemux_$BITS
-    TARGETDIR=${TOP}/${PACKAGE_DIR}/avidemux_${BITS}
-    if [ ! -e ${TARGETDIR}/platforms ]; then
-        mkdir ${TARGETDIR}/platforms;
+    TARGETDIR="${PACKAGE_DIR}/avidemux_$BITS"
+    if [ ! -e "${TARGETDIR}"/platforms ]; then
+        mkdir "${TARGETDIR}"/platforms || fail "creating platforms directory"
     fi
-    if [ ! -e ${TARGETDIR}/styles ]; then
-        mkdir ${TARGETDIR}/styles;
+    if [ ! -e "${TARGETDIR}"/styles ]; then
+        mkdir "${TARGETDIR}"/styles || fail "creating styles directory"
     fi
-    cd ${MINGW}/bin
+    cd "${MINGW}"/bin
     if [ "x${external_liba52}" = "x1" ]; then
-        cp -v liba52-*.dll $TARGETDIR
-    fi    
+        cp -v liba52-*.dll "$TARGETDIR"
+    fi
     if [ "x${external_libass}" = "x1" ]; then
-        cp -v libass-*.dll $TARGETDIR
+        cp -v libass-*.dll "$TARGETDIR"
     fi
     if [ "x${external_libmad}" = "x1" ]; then
-        cp -v libmad-*.dll $TARGETDIR
+        cp -v libmad-*.dll "$TARGETDIR"
     fi
     cp -v \
     libaom.dll \
@@ -262,32 +269,34 @@ create_release_package()
     SDL2.dll \
     xvidcore.dll \
     zlib1.dll \
-    $TARGETDIR;
-    cd $QT_HOME
+    "$TARGETDIR";
+    cd "$QT_HOME"
     cp -v \
     bin/Qt5Core.dll \
     bin/Qt5Gui.dll \
     bin/Qt5Network.dll \
     bin/Qt5Widgets.dll \
     bin/Qt5WinExtras.dll \
-    $TARGETDIR;
+    "$TARGETDIR";
     cp -v \
     plugins/platforms/qminimal.dll \
     plugins/platforms/qwindows.dll \
-    ${TARGETDIR}/platforms/;
+    "${TARGETDIR}"/platforms/;
     cp -v \
     plugins/styles/qwindowsvistastyle.dll \
-    ${TARGETDIR}/styles/;
-    mkdir ${TARGETDIR}/etc
-    cp -rvL ${MINGW}/etc/fonts ${TARGETDIR}/etc
-    cd $TARGETDIR
-    find . -name "*.dll.a" -exec rm -v '{}' \;
-    rm -Rf include
-    cd ../
+    "${TARGETDIR}"/styles/;
+    mkdir "${TARGETDIR}"/etc || fail "creating etc directory"
+    cp -rvL "${MINGW}"/etc/fonts "${TARGETDIR}"/etc
+    cd "$TARGETDIR"
+    if [ ! "x$debug" = "x1" ]; then
+        find . -name "*.dll.a" -exec rm -v '{}' \;
+        rm -Rf include
+    fi
+    cd "$PACKAGE_DIR"
     zip -r avidemux_r${BUILDDATE}_win${BITS}Qt5.zip avidemux_$BITS
     rm -Rf avidemux_$BITS
-    cd $TOP
-    echo "Avidemux Windows package generated as ${TOP}/${PACKAGE_DIR}/avidemux_r${BUILDDATE}_win${BITS}Qt5.zip"
+    popd > /dev/null
+    echo "Avidemux Windows package generated as \"${PACKAGE_DIR}/avidemux_r${BUILDDATE}_win${BITS}Qt5.zip\""
 }
 
 # Options handling
@@ -372,49 +381,49 @@ setupEnv
 
 echo "** Bootstrapping Avidemux **"
 if [ "x$author_setup" = "x1" ]; then
-    rm -Rf ${MINGWDEV}/*
+    rm -Rf "${MINGWDEV}"/*
 fi
 if [ -e "$INSTALL_DIR" -a "x$do_core" = "x1" -a "x$do_qt" = "x1" ]; then
     rm -Rf "$INSTALL_DIR";
 fi
-mkdir -p $INSTALL_DIR
-echo "Top dir : $TOP"
+mkdir -p "$INSTALL_DIR"
+echo "Build top dir : $BUILDTOP"
 
 # Build and install Avidemux components to the destination dir
 
 if [ "x$do_core" = "x1" ]; then 
     echo "** CORE **"
-    Process buildMingwCore-${ARCH} ../avidemux_core
+    Process buildMingwCore-${ARCH} "${SRCTOP}"/avidemux_core "-DCMAKE_CROSS_PREFIX=${CROSS_PREFIX}"
 fi
 
 if [ "x$do_qt" = "x1" ]; then
     echo "** QT **"
-    Process buildMingwQt5-${ARCH} ../avidemux/qt4 "-DQT_HOME:STRING=$QT_HOME -DENABLE_QT5=true"
+    Process buildMingwQt5-${ARCH} "${SRCTOP}"/avidemux/qt4 "-DQT_HOME:STRING=${QT_HOME} -DENABLE_QT5=true"
 fi
 
 if [ "x$do_cli" = "x1" ]; then
     echo "** CLI **"
-    Process buildMingwCli-${ARCH} ../avidemux/cli
+    Process buildMingwCli-${ARCH} "${SRCTOP}"/avidemux/cli
 fi
 
 if [ "x$do_plugins" = "x1" ]; then
     echo "** Plugins **"
-    Process buildMingwPluginsCommon-${ARCH} ../avidemux_plugins "-DPLUGIN_UI=COMMON $EXTRA_CMAKE_DEFS"
+    Process buildMingwPluginsCommon-${ARCH} "${SRCTOP}"/avidemux_plugins "-DPLUGIN_UI=COMMON $EXTRA_CMAKE_DEFS"
 fi
 
 if [ "x$do_plugins" = "x1" -a "x$do_qt" = "x1" ]; then
     echo "** Plugins Qt **"
-    Process buildMingwPluginsQt5-${ARCH} ../avidemux_plugins "-DPLUGIN_UI=QT4 -DQT_HOME:STRING=$QT_HOME -DENABLE_QT5=true $EXTRA_CMAKE_DEFS"
+    Process buildMingwPluginsQt5-${ARCH} "${SRCTOP}"/avidemux_plugins "-DPLUGIN_UI=QT4 -DQT_HOME:STRING=$QT_HOME -DENABLE_QT5=true $EXTRA_CMAKE_DEFS"
 fi
 
 if [ "x$do_plugins" = "x1" -a "x$do_cli" = "x1" ]; then
     echo "** Plugins CLI **"
-    Process buildMingwPluginsCli-${ARCH} ../avidemux_plugins "-DPLUGIN_UI=CLI $EXTRA_CMAKE_DEFS"
+    Process buildMingwPluginsCli-${ARCH} "${SRCTOP}"/avidemux_plugins "-DPLUGIN_UI=CLI $EXTRA_CMAKE_DEFS"
 fi
 
 if [ "x$do_plugins" = "x1" ]; then
     echo "** Plugins Settings **"
-    Process buildMingwPluginsSettings-${ARCH} ../avidemux_plugins "-DPLUGIN_UI=SETTINGS $EXTRA_CMAKE_DEFS"
+    Process buildMingwPluginsSettings-${ARCH} "${SRCTOP}"/avidemux_plugins "-DPLUGIN_UI=SETTINGS $EXTRA_CMAKE_DEFS"
 fi
 
 # Create a ZIP archive with avidemux and all the required libs
@@ -424,5 +433,3 @@ if [ "x$do_release_pkg" = "x1" ]; then
 fi
 
 echo "** All done **"
-cd $TOP
-echo "** ALL DONE **"
